[CCS-2986] move Edit Metadata to revisions dropdown
