+++ frontend/src/app/moduleDisplay.test.tsx
@@ -61,106 +61,4 @@ describe('ModuleDisplay tests', () => {
         const dataListItemRow = wrapper.find(DataListItemRow);
         expect(dataListItemRow.exists()).toBe(true)
     });
-
-    // it('should render a form group', () => {
-    //     const wrapper = mount(<ModuleDisplay {...props} />);
-    //     const formGroup = wrapper.find(FormGroup);
-    //     expect(formGroup.exists()).toBe(true)
-    // });
-
-    it('should render a Modal', () => {
-        const wrapper = mount(<ModuleDisplay {...props} />);
-        const modal = wrapper.find(Modal);
-        expect(modal.exists()).toBe(true)
-    });
-
-    // it('should render a FormSelect Element', () => {
-    //     const wrapper = mount(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test", login: true})
-    //     const formSelect = wrapper.find(FormSelect);
-    //     expect(formSelect.exists()).toBe(true)
-    // });
-
-    // it('should render a text input', () => {
-    //     const wrapper = mount(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const textInput = wrapper.find(TextInput);
-    //     expect(textInput.exists()).toBe(true)
-    // });
-
-    // it('test onChangeProduct function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.onChangeProduct("testProduct", event)).toMatchSnapshot();
-    // });
-
-    // it('test onChangeVersion function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.onChangeVersion("testVersion", event)).toMatchSnapshot();
-    // });
-
-    // it('test onChangeUsecase function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.onChangeUsecase("testUsecase", event)).toMatchSnapshot();
-    // });
-
-    // it('test handleURLInput function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.handleURLInput("testURL")).toMatchSnapshot();
-    // });
-
-    // it('test productUrlExist function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.productUrlExist("testURL")).toMatchSnapshot();
-    // });
-
-    // it('test getProductUrl function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.getProductUrl("testurl")).toMatchSnapshot();
-    // });
-
-    // it('test fetchProductVersionDetails function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     wrapper.setState({moduleName: "test", modulePath: "modules/test"})
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.fetchProductVersionDetails()).toMatchSnapshot();
-    // });
-    
-    // it('test handleModalToggle function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.handleModalToggle()).toMatchSnapshot();
-    // });
-
-    // it('test loginRedirect function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.loginRedirect).toMatchSnapshot();
-    // });
-
-    // it('test checkAuth function', () => {
-    //     const wrapper = renderer.create(<ModuleDisplay {...props} />);
-    //     const inst = wrapper.getInstance();
-    //     expect(inst.checkAuth).toMatchSnapshot();
-    // });
-
-    // it('should handle state changes for login', () => {
-    //     const wrapper = shallow(<ModuleDisplay {...props} />)
-
-    //     expect(wrapper.state('login')).toBe(false)
-    //     wrapper.setState({ 'login': true })
-    //     expect(wrapper.state('login')).toBe(true)
-    // });
-
 });
